+++ src/cvservice/api/get_cvmodel_api.go
@@ -2,8 +2,9 @@ package api
 
 import (
 	"context/application/cvservice/models"
-	"context/application/cvservice/services"
+	"context/application/cvservice/repositories"
 	"encoding/json"
+	"io/ioutil"
 	"net/http"
 )
 
@@ -11,7 +12,7 @@ func ReadCVModel(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=ascii")
 	w.Header().Set("Access-Control-Allow-Origin", "*")
 	w.Header().Set("Access-Control-Allow-Headers", "Content-Type,access-control-allow-origin, access-control-allow-headers")
-	cvEntities, infoEntities := services.FindAllCVModel()
+	cvEntities, infoEntities := repositories.FindAllCVModel()
 	results := []models.CVModel{}
 	for i := 0; i < len(cvEntities); i++ {
 		content := []models.Info{}
@@ -32,4 +33,12 @@ func ReadCVModel(w http.ResponseWriter, r *http.Request) {
 		results = append(results, cvModel)
 	}
 	json.NewEncoder(w).Encode(results)
+}
+
+func UpdateCV(w http.ResponseWriter, r *http.Request) {
+	reqBody, _ := ioutil.ReadAll(r.Body)
+	var cv []models.CVModel
+	json.Unmarshal(reqBody, &cv)
+	repositories.UpdateCV(cv)
+	json.NewEncoder(w).Encode(cv)
 }
