+++ src/cvservice/entities/go.mod
@@ -1,0 +1,3 @@
+module entities
+
+go 1.16
