+++ src/cvservice/handlers/handlers.go
@@ -4,7 +4,6 @@ package handlers
 import (
 	"context/application/cvservice/api"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
 
@@ -13,15 +12,9 @@ import (
 
 func HandleRequest() {
 	myRouter := mux.NewRouter().StrictSlash(true)
-	myRouter.Use(CORS)
-	myRouter.HandleFunc("/cvmodel", api.UpdateCV).Methods("POST")
+	// myRouter.HandleFunc("/cvmodel").Methods("POST")
 	myRouter.HandleFunc("/cvmodel", api.UpdateCV).Methods("PUT")
 	myRouter.HandleFunc("/cvmodel", api.ReadCVModel).Methods("GET")
 	fmt.Println("GO SERVER IS RUNNING. . . ")
 	log.Fatal(http.ListenAndServe(":8080", myRouter))
-}
-
-func createCVModel(w http.ResponseWriter, r *http.Request) {
-	reqBody, _ := ioutil.ReadAll(r.Body)
-	fmt.Fprintf(w, "%+v", string(reqBody))
 }
