--- src/cvservice/services/find_cvmodel_service.go
@@ -1,10 +1,0 @@
-package services
-
-import (
-	"context/application/cvservice/entities"
-	"context/application/cvservice/repositories"
-)
-
-func FindAllCVModel() (entities.CVEntities, entities.InfoEntities) {
-	return repositories.FindAllCVModel()
-}
