--- src/cvservice/services/update_cv_service.go
@@ -1,10 +1,0 @@
-package services
-
-import (
-	"context/application/cvservice/models"
-	"context/application/cvservice/repositories"
-)
-
-func UpdateCV(cv []models.CVModel) {
-	repositories.UpdateCV(cv)
-}
