--- src/cvservice/utils/convert.go
@@ -1,1 +1,0 @@
-package utils
