+++ src/go.mod
@@ -2,8 +2,14 @@ module context/application
 
 go 1.16
 
-require (
-	github.com/go-sql-driver/mysql v1.6.0 // indirect
-	github.com/gorilla/mux v1.8.0 // indirect
-	github.com/lib/pq v1.10.2 // indirect
-)
+replace context/application/cvservice/handlers => /cvservice/handlers
+
+replace context/application/cvservice/api => /cvservice/api
+
+replace context/application/cvservice/models => /cvservice/models
+
+replace context/application/cvservice/repositories => /cvservice/repositories
+
+replace context/application/cvservice/entities => /cvservice/entities
+
+require context/application/cvservice/handlers v0.0.0-00010101000000-000000000000
